% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantemp_barcode.R
\name{cantemp_barcode}
\alias{cantemp_barcode}
\title{Create a barcode plot from a time-series of temperature data}
\usage{
cantemp_barcode(data, scale_col_low = "#053061",
  scale_col_mid = "white", scale_col_high = "#67001f",
  scale_col_na = "grey50", scale_name = NULL, scale_breaks = 5,
  scale_digits = 1, x_breaks = waiver())
}
\arguments{
\item{data}{A \code{data.frame} returned by \code{\link{cantemp_fetch}},
subset to a single climate station}

\item{scale_col_low}{Color for low end of temperature gradient (see
\code{\link[ggplot2]{scale_fill_gradient2}})}

\item{scale_col_mid}{Color for middle of temperature gradient (see
\code{\link[ggplot2]{scale_fill_gradient2}})}

\item{scale_col_high}{Color for high end of temperature gradient (see
\code{\link[ggplot2]{scale_fill_gradient2}})}

\item{scale_col_na}{Color for missing values of temperature (see
\code{\link[ggplot2]{scale_fill_gradient2}})}

\item{scale_name}{Name to print above color scale}

\item{scale_breaks}{Number of breaks in color scale}

\item{scale_digits}{Number of digits in color scale temperature labels}

\item{x_breaks}{Breaks for the x-axis (see 'breaks' argument of
\code{\link[ggplot2]{scale_x_continuous}})}
}
\value{
A barcode plot
}
\description{
Create a barcode plot from a time-series of temperature data.
}
\examples{
temp_annual <- cantemp_fetch(interval = "annual")
temp_toronto <- subset(temp_annual, station == "TORONTO")

cantemp_barcode(temp_toronto)

}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
