% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantempr_fetch.R
\name{cantemp_fetch}
\alias{cantemp_fetch}
\title{Fetch homogenized temperatures for 339 Canadian Climate Stations}
\usage{
cantemp_fetch(interval = "annual")
}
\arguments{
\item{interval}{The interval of interest. One of \code{"monthly"} (monthly
intervals), \code{"seasonal"} (seasonal intervals), \code{"annual"} (annual
intervals), or \code{"all"} (all intervals combined).}
}
\value{
A \code{data.frame} with homogenized temperature data for 339
Canadian Climate stations at the specified \code{interval}.

describe{
item{prov}{Two-letter abbreviation for province}
item{station}{Name of climate station}
item{stnid}{Climate station identifier}
item{year}{Observation year}
item{period}{Observation period (either a month, season, or "Annual", as
specified in argument \code{interval})}
item{temp}{Observed mean temperature in degrees celcius}
item{flag}{Flag giving information about temperature data (\code{M}: missing
value, \code{E}: estimated during archiving process, \code{a}: adjusted for
homogeneity, \code{NA}: original value)}
}
}
\description{
Fetch a tidy \code{data.frame} with homogenized temperatures for 339 Canadian
Climate Stations, optionally subset to a given interval (monthly, seasonal,
or annual).
}
\examples{
# temperatures for monthly intervals
clim_month <- cantemp_fetch(interval = "monthly")

# temperatures for seasonal intervals
clim_season <- cantemp_fetch(interval = "seasonal")

# temperatures for annual intervals
clim_annual <- cantemp_fetch(interval = "annual")

# temperatures for all intervals
clim_all <- cantemp_fetch(interval = "all")

}
\references{
Vincent, L. A., X. L. Wang, E. J. Milewska, H. Wan, F. Yang, and V. Swail
(2012). A second generation of homogenized Canadian monthly surface air
temperature for climate trend analysis. Journal of Geophysical Research
117(D18110). \url{https://doi.org/10.1029/2012JD017859}
}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
