% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantempr-package.R
\docType{package}
\name{cantempr-package}
\alias{cantempr-package}
\title{\pkg{cantempr}}
\description{
An R package to access homogenized temperature data for 338 Canadian climate
stations. These data come from
\href{https://doi.org/10.1029/2012JD017859}{Vincent et al. (2012)}, and are
available from
\href{https://www.canada.ca/en/environment-climate-change/services/climate-change/science-research-data/climate-trends-variability/adjusted-homogenized-canadian-data/surface-air-temperature-access.html}{Environment
and Climate Change Canada}.
}
\references{
Vincent, L. A., X. L. Wang, E. J. Milewska, H. Wan, F. Yang, and V. Swail
(2012). A second generation of homogenized Canadian monthly surface air
temperature for climate trend analysis. Journal of Geophysical Research
117(D18110). \url{https://doi.org/10.1029/2012JD017859}
}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
